+++ src/main/java/com/github/gianlucanitti/javaexpreval/Function.java
@@ -1,6 +1,7 @@
 package com.github.gianlucanitti.javaexpreval;
 
 import java.io.PrintWriter;
+import java.math.BigDecimal;
 
 /**
  * Defines a generic function, with a name a number of arguments and a read-only flag.
@@ -38,7 +39,7 @@ public abstract class Function {
 
     /**
      * Evaluates this function for the specified arguments in the specified context.
-     * This is a public wrapper for {@link #evalFunction(double[], ExpressionContext, PrintWriter)}: ensures argument count is as expected, then calls it.
+     * This is a public wrapper for {@link #evalFunction(BigDecimal[], ExpressionContext, PrintWriter)}: ensures argument count is as expected, then calls it.
      * @param args The values to pass to the function.
      * @param context The context this function must be evaluated into
      * (this is necessary, for example, if this is a {@link CustomFunction} defined as an expression with references to variables or other functions).
@@ -48,7 +49,7 @@ public abstract class Function {
      * @throws UndefinedException if the number of arguments supplied is different than the number of arguments expected by this function
      * (see {@link #getArgCount()}) or if the evaluation process throws an {@link UndefinedException} itself.
      */
-    public final double eval(double[] args, ExpressionContext context, PrintWriter logWriter) throws UndefinedException{
+    public final BigDecimal eval(BigDecimal[] args, ExpressionContext context, PrintWriter logWriter) throws UndefinedException{
         if(args.length != getArgCount())
             throw new UndefinedException(getName(), args.length);
         return evalFunction(args, context, logWriter);
@@ -62,14 +63,14 @@ public abstract class Function {
     /**
      * Evaluates this function for the specified arguments in the specified context.
      * This must be overridden by subclasses and doesn't ensure that the number of arguments is correct
-     * (this is done by the public wrapper {@link #eval(double[], ExpressionContext, PrintWriter)}).
+     * (this is done by the public wrapper {@link #eval(BigDecimal[], ExpressionContext, PrintWriter)}).
      * @param args The values to pass to the function.
      * @param context The context this function must be evaluated into.
      * @param logWriter The {@link java.io.Writer} to write evaluation steps onto.
      * @return The result of the evaluation.
      * @throws UndefinedException If the function can't be evaluated because its definition references undefined variables or functions.
      */
-    protected abstract double evalFunction(double[] args, ExpressionContext context, PrintWriter logWriter) throws UndefinedException;
+    protected abstract BigDecimal evalFunction(BigDecimal[] args, ExpressionContext context, PrintWriter logWriter) throws UndefinedException;
 
     /**
      * Compares this function with another one.
